src/requirements.txt
boto3>=1.28.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0